Project_Starter_Files-Building_a_Metrics_Dashboard/app/backend/requirements.txt
Flask==3.1.*
requests==2.32.*
pylint==3.3.*
gunicorn==23.0.*
pymongo==4.11.*
Flask-PyMongo==3.0.*
prometheus-client==0.21.*
prometheus-flask-exporter==0.23.*
environs==14.1.*
opentelemetry-distro
jaeger-client